moved validation to spotifyAuthorization.js
