+++ src/utils/spotifyApiCalls.js
@@ -1,4 +1,4 @@
-import { redirectToAuthCodeFlow, getAccessToken, getRefreshToken } from "./spotifyAuthorization";
+import { redirectToAuthCodeFlow, getAccessToken } from "./spotifyAuthorization";
 
 export async function fetchAccessTokenForSearching() {
   try {
@@ -64,23 +64,11 @@ export async function fetchArtistTopTracks(name) {
 
 export async function createPlaylist(playlistName, trackUris) {
   try {
-    const code = new URLSearchParams(window.location.search).get("code");
-    let validatedToken = await validateAccessToken()
-    if (!validatedToken) {
-      if (!code) {
-        await redirectToAuthCodeFlow();
-        return false;
-      }
-      validatedToken = await getAccessToken(code);
-      if (!validatedToken) {
-        await redirectToAuthCodeFlow();
-        return false;
-      }
-      // console.error('Token validation failed.');
-      // return; // to stop executing if validation failed
-    };
-
-    const user = await fetchUser(validatedToken);
+    let accessToken = localStorage.getItem('access_token');
+    if (!accessToken) {
+      accessToken = getAccessToken();
+    }
+    const user = await fetchUser(accessToken);
 
     const response = await fetch(
       `https://api.spotify.com/v1/users/${user.id}/playlists`,
@@ -88,7 +76,7 @@ export async function createPlaylist(playlistName, trackUris) {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
-          "authorization": "Bearer " + validatedToken,
+          "authorization": "Bearer " + accessToken,
         },
         body: JSON.stringify({
           name: playlistName,
@@ -105,22 +93,11 @@ export async function createPlaylist(playlistName, trackUris) {
 
     const data = await response.json();
     const playlistId = data.id
-    await addTracksToPlaylist(playlistId, validatedToken, trackUris)
+    await addTracksToPlaylist(playlistId, accessToken, trackUris)
     return true;
   } catch (error) {
     console.error("Error:", error);
   }
-}
-
-export async function validateAccessToken() {
-  let accessToken = localStorage.getItem('access_token');
-
-  if (isTokenExpired()) {
-    console.log('check expiry')
-    accessToken = await getRefreshToken();
-  }
-
-  return accessToken;
 }
 
 export async function fetchUser(token) {
@@ -135,11 +112,6 @@ export async function fetchUser(token) {
   // saving current user for setting user state object after playlist creation
   localStorage.setItem('current_user', JSON.stringify(data));
   return data;
-}
-
-function isTokenExpired() {
-  const expirationTime = parseInt(localStorage.getItem("expires_in"), 10);
-  return Date.now() > expirationTime;
 }
 
 async function addTracksToPlaylist(playlistId, accessToken, trackUris) {
