+++ src/utils/spotifyAuthorization.js
@@ -39,7 +39,18 @@ async function generateCodeChallenge(verifier) {
 
 // Getting Access Tokens
 
-export async function getAccessToken(code) {
+export async function getAccessToken() {
+  const code = new URLSearchParams(window.location.search).get("code");
+    if (!code) {
+      await redirectToAuthCodeFlow();
+      return false;
+    }
+  //if there already is a AT refresh it
+  const localAccessToken = await validateAccessToken();
+  if (localAccessToken) {
+    return localAccessToken;
+  }
+
   const verifier = localStorage.getItem("verifier");
   if (!verifier) {
     console.error("Verifier is missing. Redirecting to authorization flow.");
@@ -77,6 +88,22 @@ export async function getAccessToken(code) {
     console.error("Error getting access token:", error);
     return null;
   }
+}
+
+async function validateAccessToken() {
+  let accessToken = localStorage.getItem('access_token');
+
+  if (isTokenExpired()) {
+    console.log('check expiry')
+    accessToken = await getRefreshToken();
+  }
+
+  return accessToken;
+}
+
+function isTokenExpired() {
+  const expirationTime = parseInt(localStorage.getItem("expires_in"), 10);
+  return Date.now() > expirationTime;
 }
 
 export async function getRefreshToken() {
